refactor: drop part of the old hooks
